requirements.txt
Django==3.2.11
djangorestframework==3.12.4
Faker==8.14.0